Crate rollenspielsache

The Rollenspielsache is a set of tools for managing tabletop RPG games. The goal is to handle the mechanics seamlessly, allowing you to participate in or run a tabletop encounter without getting distracted by rule clarifications.

This library is the core API defining all the entities involved and their relationships.

Types

Roll

RollResult

Re-exports

pub use interface::*;

pub use dice::interface::*;

Modules

character
Characters - PCs, NPCs, drones.

dice
Dice is responsible for executing dice rolls.

error
Library error type.

game_system
Supported RPG-specific rules.

interface
Crate-level C-compatible exports

modifier
All the different types of modifier.

redis
Working with Redis datatypes.

skill
skill.rs is concerned with skills and abilities.